requirements.txt
telebot
validators
requests
yt_dlp
python-dotenv